refactor: refactor necessary for migration utility
